Split log reader into separate base class
